extracted spec for api integration tests
