+++ prinz/src/test/scala/pl/touk/nussknacker/prinz/UnitTest.scala
@@ -1,8 +1,11 @@
 package pl.touk.nussknacker.prinz
 
-import org.scalatest.{GivenWhenThen, Inside, Inspectors, OptionValues}
-import org.scalatest.flatspec.AnyFlatSpec
+import org.scalatest.{Inside, Inspectors, OptionValues}
+import org.scalatest.flatspec.AnyFlatSpecLike
 import org.scalatest.matchers.should
 
-abstract class UnitTest extends AnyFlatSpec with should.Matchers with
-  OptionValues with Inside with Inspectors
+trait UnitTest extends AnyFlatSpecLike
+  with should.Matchers
+  with OptionValues
+  with Inside
+  with Inspectors
